packages/agents/alert/pyproject.toml
[project]
name = "alert"
version = "0.1.0"
description = "天気データ分析AIエージェント - 異常検知エージェント（Step 3）"
requires-python = ">=3.12"
dependencies = [
    "strands-agents>=0.1",
    "shared",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/alert"]
